requirements.txt
google-adk>=1.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.35.0
vertexai>=0.0.1
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
numpy>=1.24.0  # Para cálculos de similitud coseno
scikit-learn>=1.0.0  # Para procesamiento de vectores 